Fix attachment realtime reconciliation regressions
